# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Product categories
  - name: products
    description: Product operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /categories/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get a category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Category not found
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, name_asc]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/highlighted products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/new-arrivals:
    get:
      operationId: getNewArrivals
      tags: [products]
      summary: Get newest products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: New arrivals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
  /products/summary:
    get:
      operationId: getProductsSummary
      tags: [products]
      summary: Get store summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - imageUrl
        - productCount
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - images
        - categoryId
        - categoryName
        - sizes
        - colors
        - inStock
        - featured
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - products
        - total
        - hasMore
    StoreSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        priceRange:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
          required:
            - min
            - max
        featuredCount:
          type: integer
      required:
        - totalProducts
        - totalCategories
        - priceRange
        - featuredCount