# lib/api-zod/src/generated/types/productListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface ProductListResponse {
  products: Product[];
  total: number;
  hasMore: boolean;
}