# lib/api-zod/src/generated/types/storeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreSummaryPriceRange } from "./storeSummaryPriceRange";

export interface StoreSummary {
  totalProducts: number;
  totalCategories: number;
  priceRange: StoreSummaryPriceRange;
  featuredCount: number;
}